requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
async-timeout==4.0.3
attrs==25.4.0
beautifulsoup4==4.14.3
cachetools==6.2.3
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
dateparser==1.0.0
defusedxml==0.7.1
discord-notify==1.0.0
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.127.0
filetype==1.2.0
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
idna==2.10
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.81
langchain-google-genai==2.0.10
langchain-openai==0.3.35
langchain-pinecone==0.2.13
langchain-text-splitters==0.3.11
langsmith==0.4.37
marshmallow==3.26.2
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.0.1
oauthlib==3.3.1
openai==2.14.0
orjson==3.11.5
packaging==24.2
pandas==2.3.3
pinecone==7.3.0
pinecone-client==6.0.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
playsound==1.2.2
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==5.8.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
pyparsing==3.2.5
pypdf==6.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-pushover==0.4
pytz==2021.1
PyYAML==5.4.1
RapidFuzz==3.13.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9.1
simsimd==6.5.12
six==1.15.0
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.45
srt==3.5.3
starlette==0.49.3
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==2.1
uritemplate==4.2.0
urllib3==1.26.4
uvicorn==0.39.0
yarl==1.22.0
youtube-transcript-api==1.2.3
yt-dlp==2025.10.14
zstandard==0.25.0
